Refactoring della struttura delle directory
