+++ components/BookCarousel.tsx
